+++ tests/test_table_xlsx.py
@@ -66,4 +66,19 @@ class TestXLSX(agate.AgateTestCase):
         self.assertColumnTypes(table, [agate.Date])
         self.assertRows(table, [
             [datetime.date(1899, 12, 31)]
+
+    def test_empty(self):
+        table = agate.Table.from_xlsx('examples/test_empty.xlsx')
+
+        self.assertColumnNames(table, [])
+        self.assertColumnTypes(table, [])
+        self.assertRows(table, [])
+
+    def test_numeric_column_name(self):
+        table = agate.Table.from_xlsx('examples/test_numeric_column_name.xlsx')
+
+        self.assertColumnNames(table, ['Country', '2013'])
+        self.assertColumnTypes(table, [agate.Text, agate.Number])
+        self.assertRows(table, [
+            ['Canada', 35160000]
         ])