+++ CHANGELOG.rst
@@ -2,6 +2,8 @@
 -----
 
 * Fix bug in handling ambiguous dates in XLS. (#9)
+* Fix bug in handling an empty XLS.
+* Fix bug in handling non-string column names in XLSX.
 
 0.2.0
 -----